# api/openapi.yaml
openapi: 3.1.0
info:
  title: sealift
  version: "1"
  description: >
    Scans a pnpm project for CVEs, ranks newer dependency versions, and
    builds an archive for an air-gapped Nexus import.
servers:
  - url: /api
...
paths:
...
  /projects/{projectId}/analyses:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      operationId: queueAnalysis
      tags:
        - analyses
      summary: Queue an analysis for the project
      responses:
        "202":
          description: Analysis queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        default:
          $ref: "#/components/responses/Problem"
  /projects/{projectId}/analyses/{analysisId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/AnalysisId"
    get:
      operationId: getAnalysis
      tags:
        - analyses
      summary: Get an analysis' status and results
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteAnalysis
      tags:
        - analyses
      summary: Delete an analysis
      responses:
        "204":
          description: Deleted
        default:
          $ref: "#/components/responses/Problem"
  /projects/{projectId}/analyses/{analysisId}/log:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/AnalysisId"
    get:
      operationId: getAnalysisLog
      tags:
        - analyses
      summary: Get an analysis' full log
      responses:
        "200":
          description: Plain text log
          content:
            text/plain:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Problem"
  /projects/{projectId}/analyses/{analysisId}/cancel:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/AnalysisId"
    post:
      operationId: cancelAnalysis
      tags:
        - analyses
      summary: Cancel a running or queued analysis
      responses:
        "200":
          description: Analysis after the cancel request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        default:
          $ref: "#/components/responses/Problem"
...
  /settings:
    get:
      operationId: getSettings
      tags:
        - settings
      summary: Get settings, with signatureKey masked
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        default:
          $ref: "#/components/responses/Problem"
    put:
      operationId: updateSettings
      tags:
        - settings
      summary: Update settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: Updated settings, with signatureKey masked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]+$'
    AnalysisId:
      name: analysisId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]+$'
...
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details for HTTP APIs
      properties:
        type:
          type: string
          format: uri
          description: A URI identifying the problem type
          default: about:blank
        title:
          type: string
          description: A short, human-readable summary of the problem type
        status:
          type: integer
          description: The HTTP status code repeated for clients that read the body only
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
        instance:
          type: string
          format: uri
          description: A URI identifying this specific occurrence
        errors:
          type: array
          description: Every offending entry, for a validation failure
          items:
            $ref: "#/components/schemas/ProblemDetail"
        missing:
          type: array
          description: Names of what a tools-missing problem needs and does not have yet
          items:
            type: string
      required:
        - title
        - status
    ProblemDetail:
      type: object
      description: One offending entry inside a validation failure's errors list
      properties:
        field:
          type: string
          description: package.json field, such as dependencies or pnpm.overrides
        name:
          type: string
          description: Dependency name, empty for a field-level problem
        value:
          type: string
        reason:
          type: string
      required:
        - field
        - reason
    State:
      type: string
      description: State of a job and of the directory it writes
      enum:
        - queued
        - running
        - done
        - failed
        - cancelled
        - interrupted
    Target:
      type: object
      description: The system a resolution or an export is prepared for
      properties:
        os:
          type: string
          example: linux
        cpu:
          type: string
          example: x64
        libc:
          type: string
          example: glibc
        node:
          type: string
          example: 22.17.1
        pnpmVer:
          type: string
          example: 10.34.5
      required:
        - os
        - cpu
        - libc
        - node
        - pnpmVer
    Settings:
      type: object
      description: Every value the interface can change.
      properties:
        target:
          $ref: "#/components/schemas/Target"
        signatureKey:
          type: string
          description: Masked on read, never logged
        minReleaseAgeDays:
          type: integer
        resolveParallelism:
          type: integer
        downloadParallelism:
          type: integer
      required:
        - target
        - signatureKey
        - minReleaseAgeDays
        - resolveParallelism
        - downloadParallelism
...
    Analysis:
      type: object
      description: One analysis job and its state.
      properties:
        id:
          type: string
        projectId:
          type: string
        state:
          $ref: "#/components/schemas/State"
        createdAt:
          type: string
          format: date-time
        result:
          $ref: "#/components/schemas/AnalysisResult"
        trivyVersion:
          type: string
          description: Trivy version the analysis used, absent when it did not record one
        trivyDbDate:
          type: string
          format: date-time
          description: Vulnerability database date the analysis used, absent when it did not record one
        pnpmVersion:
          type: string
          description: pnpm version the analysis used, absent when it did not record one
        failure:
          $ref: "#/components/schemas/Failure"
      required:
        - id
        - projectId
        - state
        - createdAt
    Failure:
      type: object
      description: Which step failed and why
      properties:
        step:
          type: string
        message:
          type: string
      required:
        - step
        - message
    Vector:
      type: array
      description: CVE counts by severity, critical to unknown
      items:
        type: integer
      minItems: 5
      maxItems: 5
    Signal:
      type: object
      description: One reason a candidate is blocked or worth a second look
      properties:
        name:
          type: string
        evidence:
          type: string
        blocking:
          type: boolean
      required:
        - name
        - evidence
        - blocking
    Candidate:
      type: object
      description: One newer version of a dependency, ranked against the current one
      properties:
        version:
          type: string
        vector:
          $ref: "#/components/schemas/Vector"
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        key:
          type: boolean
          description: Whether this candidate was resolved and scanned before the rest
        resolved:
          type: boolean
        published:
          type: string
          format: date-time
          description: Empty when the registry has no publication time
      required:
        - version
        - vector
        - signals
        - key
        - resolved
    DependencyResult:
      type: object
      description: One direct dependency's current CVEs and its ranked candidates
      properties:
        name:
          type: string
        current:
          type: string
        best:
          type: string
          description: Empty when no candidate improves on the current version
        vector:
          $ref: "#/components/schemas/Vector"
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
      required:
        - name
        - current
        - vector
        - candidates
    AnalysisResult:
      type: object
      description: What an analysis writes to candidates.json and ranking.json
      properties:
        target:
          $ref: "#/components/schemas/Target"
        before:
          $ref: "#/components/schemas/Vector"
        after:
          allOf:
            - $ref: "#/components/schemas/Vector"
          nullable: true
          x-omitempty: false
          description: Null when the combined check (step 9) never measured it
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/DependencyResult"
        warnings:
          type: array
          items:
            type: string
      required:
        - target
        - before
        - dependencies
        - warnings
...